Create LanguageGenerationExample.lg
Here I worked with Language Generation Files in Bot Framework Composer. In them we added the translation text to the responsive LG files (en-us, de-de). Also there was an issue with the button texts, which where not displayed in the dialog-files in the project structure of composer, so we added the templates for the button texts in the LG Files.