Create AdaptiveCard
AdaptiveCards were used frequently in this projects. We used the data from the API to get to the minigames. We then read out the variable x (line 47, line 55) to direct to the specific minigame dialog. In line 18 there is an example for a simple if else query to adjust the text.